Remove the back/next buttons instead of disabling them
Fixes #29

I believe this was added before `touch-action` was added to the CSS.
Without `touch-action` set, repeatedly pressing back could cause the
browser to zoom if there wasn't a button to hit. Now the page naturally
ignores double taps, there is no value in disabling the button, and it
instead just disappears entirely.